# Inline image content hash computation