--- a/src/nbl/asset/interchange/CGLILoader.cpp
+++ b/src/nbl/asset/interchange/CGLILoader.cpp
@@ -22,6 +22,7 @@ SOFTWARE.
 
 #ifdef _NBL_COMPILE_WITH_GLI_LOADER_
 
+#include "nbl/asset/interchange/CImageHasher.h"
 #include "nbl/asset/interchange/IImageAssetHandlerBase.h"
 
 #ifdef _NBL_COMPILE_WITH_GLI_
@@ -187,24 +188,32 @@ namespace nbl
 			};
 
 			uint64_t tmpDataSizePerRegionSum = {};
+			nbl::asset::CImageHasher contentHasher(imageInfo);
+
 			for (uint16_t mipLevel = 0; mipLevel < imageInfo.mipLevels; ++mipLevel)
 			{
 				const auto layerSize = getFullSizeOfLayer(mipLevel);
+				const auto mipOffset = (mipLevel * imageInfo.arrayLayers);
+
 				for (uint16_t layer = 0; layer < imageInfo.arrayLayers; ++layer)
 				{
 					const auto layersData = getCurrentGliLayerAndFace(layer);
 					const auto gliLayer = layersData.first;
 					const auto gliFace = layersData.second;
+					
+					const auto pOffset = mipOffset + layer;
 
-					assignGLIDataToRegion((reinterpret_cast<uint8_t*>(data) + tmpDataSizePerRegionSum + (layer * layerSize)), texture, gliLayer, gliFace, mipLevel, layerSize);
+					auto regionData = (reinterpret_cast<uint8_t*>(data) + tmpDataSizePerRegionSum + (layer * layerSize));
+					assignGLIDataToRegion(regionData, texture, gliLayer, gliFace, mipLevel, layerSize);
+					
+					contentHasher.hashSeq(mipLevel, layer, regionData, layerSize);
 				}
 				tmpDataSizePerRegionSum += getFullSizeOfRegion(mipLevel);
 			}
 
 			image->setBufferAndRegions(std::move(texelBuffer), regions);
 
-			// TODO: inline hashing while reading
-			auto hash = image->computeContentHash();
+			auto hash = contentHasher.finalizeSeq();
 			image->setContentHash(hash);
 
 			ICPUImageView::SCreationParams imageViewInfo = {};