## Description
moves content hash computation into image loading loop - for JPEG, GLI (OpenGL Image), OpenEXR (_only these were marked for todo_)
extends `CImageHasher` (adds inline methods for _sequential hash calculation_)

## Testing 
passes `NBL_IMAGE_HASH_RUN_TESTS` with `Debug` configuration

## TODO list:
- [x] test performance of this pull; compare with previous solution (see [#131](https://github.com/Devsh-Graphics-Programming/Nabla-Examples-and-Tests/pull/131))

<!--
By creating this pull request into Nabla, you agree to release all your past (even from previous commits) and present contributions in the Nabla repository under the Apache 2.0 license. If you're not the sole contributor, ensure that all contributors have signed the CLA agreeing to this.
-->