@tad1 2024-07-25
### Perf Test
With [#131](https://github.com/Devsh-Graphics-Programming/Nabla-Examples-and-Tests/pull/131) `24_ColorSpaceTest` now gives time of each asset loading time.
<details><summary>calcPerf.py</summary>
<p>
this is an example script I've used to scrap results from logs

```python
import re
import os.path

testLogPath = "./Testing/Temporary/LastTest.log"
csvPath = "./Testing/res.csv"
assetNameRE = '.*Asset: \"(.*)\"'
assetLoadTimeRE = '.*Asset load time: (.*) ms'
mode = "none" #I updated this one manually


files = []
res = []
assetName = ""
loadTime = ""

with open(testLogPath, "r") as logFile:
    for line in logFile:
        ANGroup = re.search(assetNameRE, line)
        ALTGroup = re.search(assetLoadTimeRE, line)
        if(ANGroup):
            assetName = ANGroup.group(1)
            files.append(assetName)
        if(ALTGroup):
            loadTime = ALTGroup.group(1)
            res.append(loadTime)


append = False
if(os.path.exists(csvPath)):
    append = True

if(append):
    with open(csvPath, "a") as file:
        file.write(f"{", ".join([mode] + res)}\n")
else:
    with open(csvPath, "w") as file:
        file.write(f"{", ".join(["mode"]+files)}\n")
        file.write(f"{", ".join([mode] + res)}\n")
```
</p>
</details> 

How to run test:
1. Make sure you are on `FlattenRegionsStreamHashImageFilter` branch (in _Nabla Examples and Tests_ repo)
2. build `24_ColorSpaceTest`
3. in `.\examples_tests\24_ColorSpaceTest` dir run `ctest -C Debug --progress --stop-on-failure --verbose; python .\calcPerf.py`

### Results
[perfResult.csv](https://github.com/user-attachments/files/16381476/perfResult.csv) _(values are in milliseconds, lower = better)_

#### Configuration:
OS: Windows 10
CPU: Intel Core i3-9100F
Build mode: Debug

I tested on: 
- **inline** (current patch)
- **original** ([f55c7](https://github.com/Devsh-Graphics-Programming/Nabla/commit/f55c778b66fc5500e8751ff6e7bbfd5ba076b898))
- **none** (skipping hash computation - see [29cdc](https://github.com/tad1/Nabla/commit/29cdc6577a9dd3ecafa0f0a0925cbb33da8cc287))

With this patch, we remove major overhead (440 ms -> 40ms for largest JPEG; 20 ms -> 6ms earth-array.dds; 15 ms -> 5ms earth-array.ktx) for loading textures.

PNG, TGA times for inline, and original are (and should be) the same (patch affects only for JPEG, OpenEXR, and GLI formats).

Hashing now takes the minority of the asset loading time.